requirements.txt
os
logging
pytesseract
fitz
multiprocessing
re
tabula-py
time
psutil
cv2
numpy
pdfplumber
PyPDF2
gc
tiktoken
spacy
pathlib
concurrent.futures
json
unicodedata
